Django projects/Laptopshop/requirements.txt
asgiref==3.8.1
attrs==25.3.0
Django==5.2.2
djangorestframework==3.16.0
drf-yasg==1.21.10
importlib_resources==6.5.2
inflection==0.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
packaging==25.0
pillow==11.2.1
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.25.1
sqlparse==0.5.3
swagger-spec-validator==3.0.4
typing_extensions==4.14.0
tzdata==2025.2
uritemplate==4.2.0
